fixed buffer error with unstandard SEGY files

+++ src/OSEGYRev0.cpp
@@ -1,5 +1,6 @@
 #include "OSEGYRev0.hpp"
 #include "Exception.hpp"
+#include <cstdint>
 #include <cstring>
 #include <functional>
 #include <vector>
@@ -56,8 +57,9 @@ OSEGYRev0::Impl::Impl(OSEGYRev0& s, string txt_hdr)
 	zero.format_code = 1;
     if (memcmp(&sgy.common().binary_header, &zero,
 			   sizeof(CommonSEGY::BinaryHeader))) {
-        sgy.common().samp_buf.resize(sgy.common().binary_header.samp_per_tr *
-									 sgy.common().bytes_per_sample);
+        sgy.common().samp_buf.resize(static_cast<uint16_t>(
+				sgy.common().binary_header.samp_per_tr) *
+			sgy.common().bytes_per_sample);
         write_trace = [this](Trace& tr) {
             sgy.write_trace_header(tr.header());
             sgy.write_trace_samples_fix(tr);
